chore(compliance): simplify ComplianceOutput class
