+++ prowler/lib/outputs/compliance/cis/cis_gcp.py
@@ -1,6 +1,3 @@
-from csv import DictWriter
-from venv import logger
-
 from prowler.lib.check.compliance_models import ComplianceBaseModel
 from prowler.lib.outputs.compliance.cis.models import CISGCP
 from prowler.lib.outputs.compliance.compliance_output import ComplianceOutput
@@ -17,7 +14,6 @@ class GCPCIS(ComplianceOutput):
 
     Methods:
         - transform: Transforms findings into GCP CIS compliance format.
-        - batch_write_data_to_file: Writes the findings data to a CSV file in GCP CIS compliance format.
     """
 
     def transform(
@@ -99,26 +95,3 @@ class GCPCIS(ComplianceOutput):
                         Muted=False,
                     )
                     self._data.append(compliance_row)
-
-    def batch_write_data_to_file(self) -> None:
-        try:
-            if (
-                getattr(self, "_file_descriptor", None)
-                and not self._file_descriptor.closed
-                and self._data
-            ):
-                csv_writer = DictWriter(
-                    self._file_descriptor,
-                    fieldnames=[field.upper() for field in self._data[0].dict().keys()],
-                    delimiter=";",
-                )
-                csv_writer.writeheader()
-                for finding in self._data:
-                    csv_writer.writerow(
-                        {k.upper(): v for k, v in finding.dict().items()}
-                    )
-                self._file_descriptor.close()
-        except Exception as error:
-            logger.error(
-                f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
-            )
