+++ prowler/lib/outputs/compliance/compliance_output.py
@@ -1,7 +1,9 @@
+from csv import DictWriter
 from io import TextIOWrapper
 from typing import List
 
 from prowler.lib.check.compliance_models import ComplianceBaseModel
+from prowler.lib.logger import logger
 from prowler.lib.outputs.finding import Finding
 from prowler.lib.outputs.output import Output
 
@@ -44,3 +46,32 @@ class ComplianceOutput(Output):
             self.transform(findings, compliance, compliance_name)
             if create_file_descriptor:
                 self.create_file_descriptor(file_path)
+
+    def batch_write_data_to_file(self) -> None:
+        """
+        Writes the findings data to a CSV file in the specific compliance format.
+
+        Returns:
+            - None
+        """
+        try:
+            if (
+                getattr(self, "_file_descriptor", None)
+                and not self._file_descriptor.closed
+                and self._data
+            ):
+                csv_writer = DictWriter(
+                    self._file_descriptor,
+                    fieldnames=[field.upper() for field in self._data[0].dict().keys()],
+                    delimiter=";",
+                )
+                csv_writer.writeheader()
+                for finding in self._data:
+                    csv_writer.writerow(
+                        {k.upper(): v for k, v in finding.dict().items()}
+                    )
+                self._file_descriptor.close()
+        except Exception as error:
+            logger.error(
+                f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
+            )
