+++ prowler/lib/outputs/compliance/iso27001/iso27001_aws.py
@@ -1,6 +1,3 @@
-from csv import DictWriter
-
-from prowler.lib import logger
 from prowler.lib.check.compliance_models import ComplianceBaseModel
 from prowler.lib.outputs.compliance.compliance_output import ComplianceOutput
 from prowler.lib.outputs.compliance.iso27001.models import ISO27001AWS
@@ -17,7 +14,6 @@ class AWSISO27001(ComplianceOutput):
 
     Methods:
         - transform: Transforms findings into AWS ENS compliance format.
-        - batch_write_data_to_file: Writes the findings data to a CSV file in AWS ENS compliance format.
     """
 
     def transform(
@@ -83,32 +79,3 @@ class AWSISO27001(ComplianceOutput):
                         Muted=False,
                     )
                     self._data.append(compliance_row)
-
-    def batch_write_data_to_file(self) -> None:
-        """
-        Writes the findings data to a CSV file in AWS ENS compliance format.
-
-        Returns:
-            - None
-        """
-        try:
-            if (
-                getattr(self, "_file_descriptor", None)
-                and not self._file_descriptor.closed
-                and self._data
-            ):
-                csv_writer = DictWriter(
-                    self._file_descriptor,
-                    fieldnames=[field.upper() for field in self._data[0].dict().keys()],
-                    delimiter=";",
-                )
-                csv_writer.writeheader()
-                for finding in self._data:
-                    csv_writer.writerow(
-                        {k.upper(): v for k, v in finding.dict().items()}
-                    )
-                self._file_descriptor.close()
-        except Exception as error:
-            logger.error(
-                f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
-            )
