+++ prowler/lib/outputs/compliance/mitre_attack/mitre_attack_aws.py
@@ -1,6 +1,3 @@
-from csv import DictWriter
-from venv import logger
-
 from prowler.lib.check.compliance_models import ComplianceBaseModel
 from prowler.lib.outputs.compliance.compliance_output import ComplianceOutput
 from prowler.lib.outputs.compliance.mitre_attack.models import MitreAttackAWS
@@ -18,7 +15,6 @@ class AWSMitreAttack(ComplianceOutput):
 
     Methods:
         - transform: Transforms findings into AWS MITRE ATT&CK compliance format.
-        - batch_write_data_to_file: Writes the findings data to a CSV file in AWS MITRE ATT&CK compliance format.
     """
 
     def transform(
@@ -117,32 +113,3 @@ class AWSMitreAttack(ComplianceOutput):
                         Muted=False,
                     )
                     self._data.append(compliance_row)
-
-    def batch_write_data_to_file(self) -> None:
-        """
-        Writes the findings data to a CSV file in AWS MITRE ATT&CK compliance format.
-
-        Returns:
-            - None
-        """
-        try:
-            if (
-                getattr(self, "_file_descriptor", None)
-                and not self._file_descriptor.closed
-                and self._data
-            ):
-                csv_writer = DictWriter(
-                    self._file_descriptor,
-                    fieldnames=[field.upper() for field in self._data[0].dict().keys()],
-                    delimiter=";",
-                )
-                csv_writer.writeheader()
-                for finding in self._data:
-                    csv_writer.writerow(
-                        {k.upper(): v for k, v in finding.dict().items()}
-                    )
-                self._file_descriptor.close()
-        except Exception as error:
-            logger.error(
-                f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
-            )
